refactor!: rename the command type from “command” to “code”
The ERCP Basic specification currently refers about the type of a command. This
name, however, cannot be used in Rust, so doesn’t in other languages, as `type`
is a reserved keyword. I have used the name `command` instead in the code until
then, but it is confusing. Let’s use `code` instead, as this will be the
official name in the next ERCP Basic specification releases.

BREAKING CHANGE: `Command::command` is now `Command::code`.